[reland] Flip triton kernel default layout constraint to "needs_fixed_stride_order"
This is to match the default layout constraint for custom operators. By
default, Inductor should match the stride order of inputs to a triton
kernel.

IF THIS IS BREAKING YOU, PLEASE REACH OUT, especially if it's been
more than two weeks since this landed. You can flip the config locally
as a workaround.

Test Plan:
- existing tests

ghstack-source-id: 44c7628c0641590fc906042c6f33fb32324464c0
Pull Request resolved: https://github.com/pytorch/pytorch/pull/137064