[export] Add optimization passes
Pull Request resolved: https://github.com/pytorch/pytorch/pull/136191


Added an optimization pass to the swap function which removes extraneous pytrees. Currently it removes the pytree flatten/unflatten calls between modules in very specific scenarios (all the inputs of one module go into the other).

Future work can be to remove the input pytree.flatten if the inputs go directly into an unflatten, and output pytree unflatten if the outputs are directly from a pytree.flatten.



@imported-using-ghimport

Differential Revision: [D62879820](https://our.internmc.facebook.com/intern/diff/D62879820/)
ghstack-source-id: 245757117